### 2020-12-15 #418 @GLOrozcoM
epair - get data from the Environmental Protection Agency Air Quality System API
Submitting Author Name: G. Leonard Orozco-Mulfinger
Submitting Author Github Handle: @GLOrozcoM
Repository: https://github.com/GLOrozcoM/epair
Version submitted: 0.1.0
Submission type: Standard
Version accepted: TBD

---

-   The full DESCRIPTION file is found below:

```
Package: epair
Title: Grabs data from EPA API, simplifies getting pollutant data
Version: 0.1.0
Authors@R: person("G.L.", "Orozco-Mulfinger", email = "glo003@bucknell.edu", role = c("aut", "cre"))
  A package to aid the user in making queries to the EPA API site found at https://aqs.epa.gov/aqsweb/documents/data_api.
  It combines API calling methods from various web scraping packages with specific strings to retrieve data from the EPA API. It also contains
  easy to use loaded variables that help a user navigate services offered by the API and aid the user in
  determining the appropriate way to make a an API call.
  xml2 (>= 1.1.1),
  rvest (== 0.3.5),
  jsonlite (>= 1.6.1)
Suggests: testthat
```

This package assists a user in making API calls to get pollutant data from the Environmental Protection Agency's Air Quality System API.

The target audience is air quality researchers in general who are not necessarily advanced R users. These data are used by researchers from a diverse range of disciplines interested in monitoring and modeling air pollution concentrations in the US. These disciplines include statistics, environmental health, air pollution, climatology, epidemiology, economics, public health, geoscience, atmospheric science, to name a few. These data are frequently used by federal agencies, academic researchers, and industry users.

The current method of downloading these data from the AQS API requires users to build their own API calls by appending strings together. However, to accurately do this requires a deep understanding of the AQS API building system, and some facility with understanding URL construction. This package is aimed at such researchers who are interested in downloading the data but do not want to invest considerable time in learning how to develop correct API calls, as well as those who want to explore the types of data available before actually downloading the data.

To date, another package (`aqsr` by `jpkeller` on GitHub) exists fulfilling similar aims to `epair`. Both packages perform an httr call to retrieve data from the EPA’s API. A few key differences though, make `epair` stand out.

The first is that `epair` provides a substantial number of aids to R users in determining how to make an API call. `epair` does this through a comprehensive services object to help the user explore EPA API services from R. For instance, besides just listing all available services using `names(services)`, the user can check a description, available filters, endpoints associated with filters, required and optional variables, and examples for these calls. `epair` also offers the variables object. It contains helpful descriptions for the user to know more about variables within the EPA API.

A second difference lies within documentation. A full PDF manual, thorough documentation for each function, and a full testing suite using `testthat` has been created in the package for maintenance.

A final new feature within `epair` is the ability to call raw JSON and R data frames. Through `perform.call()` and `perform.call.raw()`, a user can easily get an R data frame or a raw JSON result depending on their needs.

N/A

See the following [issue](https://github.com/ropensci/software-review/issues/415) for the presubmission inquiry.

## Comments

Please note that a vignette is still a work in progress, but a full tutorial website can be found [here](https://epair.netlify.app/).

Also, that no CI is included as the package doesn't make system calls or perform any of the functions outlined in the specifications for CI.
