### 2021-01-05 #418 @maelle
Thanks @GLOrozcoM for your submission! I see a few issues need to be tackled before I look for reviewers, see below. I am happy to answer questions as needed!

### Editor checks:

I see you have a documentation website, but could you re-format it to use pkgdown? See https://devguide.ropensci.org/building.html#website

It might make sense to add some comments on how the package differs from trying to get EPA data with https://docs.ropensci.org/ropenaq (I assume it is easier to find the data you need with epair, and that epair has higher coverage).

    * Your repository does not have any CI set up. We currently recommend using GitHub Actions. See e.g. https://usethis.r-lib.org/reference/use_github_action.html and https://github.com/r-lib/actions/

    * When I try to run the tests, I get errors

	    ```r
		    Failure (test_call_construction.R:25:3): Single variable added to query properly
		    `found_query` not equal to `exp_query`.
		    1/1 mismatches
		    x[1]: "https://aqs.epa.gov/data/api/dailyData/byState?&\"37\"=37"
		    y[1]: "https://aqs.epa.gov/data/api/dailyData/byState?&state=37"

		    Error (test_call_construction.R:35:3): Multiple variables added to query properly
		    Error: l'argument "name" est manquant, avec aucune valeur par défaut
		    Backtrace:
		     1. epair::add.variables(endpoint, variable.list) test_call_construction.R:35:2

		    Error (test_call_construction.R:42:3): Variables containing spaces are accepted into a potential call
		    Error: l'argument "name" est manquant, avec aucune valeur par défaut
		    Backtrace:
		     1. epair::add.variables(base, vars) test_call_construction.R:42:2
		    ```

    * In test files now context() is no longer recommended, see https://www.tidyverse.org/blog/2020/10/testthat-3-0-0/
    * testthat and epair only needs to be loaded in tests/testthat.R, not in individual files.
    * I see you skip tests using API calls. Scott Chamberlain and I wrote a whole book dedicated to HTTP testing https://books.ropensci.org/http-testing/ The Whole Game(s) section in particular can help you see how to tackle testing, and I am happy to provide guidance as needed.

---

#### Editor comments

Some further comments

* Your package uses web scraping. Is it allowed by the EPA? E.g. see https://docs.ropensci.org/robotstxt/
* Your package depends on a particular version of rvest. Why? We can't accept such a dependency because it could make your package incompatible with other packages. Dependencies should be "on this version or higher".
* I see a few typos with `devtools::spell_check()` so I'd encourage you to run it.

---

Reviewers:
Due date:
