### 2021-01-05 #418 @maelle
I see you didn't use CI because the package doesn't perform any system call: these criteria you outline are for having CI _on all operating systems_. We require a CI setup for all packages, it's very useful. :-)
