### 2021-01-14 #418 @GLOrozcoM
Response:

Hi @maelle! I’ve made some progress and would like to point some of it out here. I’ll go in order:

1. A colleague will be working on transferring the documentation website to pkgdown. You can expect to see them address this issue in the near future.
2. I’ve added a comment about `ropenaq` to the README about this. See [commit](https://github.com/GLOrozcoM/epair/commit/7f340db722a92264738383fa00d4f6a6eb0bd678).
3. Another colleague will be managing the setup for a CI service so you can expect them to address this in the near future as well.
4. I fixed the errors you pointed out. See [commit](https://github.com/GLOrozcoM/epair/commit/9b2a68ec3f9c390ae2037d4afe0842d697694e3b).
5. I’ve removed unnecessary context(), and loads for testthat and epair. See [commit](https://github.com/GLOrozcoM/epair/commit/0af0b10f901b1e3b9d7a53d9a12978c7b7883ed1).
6. I’ve been studying your book and am currently working on developing better tests.
7. I emailed the EPA team and heard back that the kind of scraping `epair` does is acceptable.
8. Thank you for pointing the versioning issue out! At one point, `rvest` had an issue with an HTML table scraping method, and my team wanted to make sure we signaled what version this issue was found in. The recent release of `rvest` fixed this issue, and given your feedback, I’ve modified the description to read “on this version or higher”. See [commit](https://github.com/GLOrozcoM/epair/commit/198bd624147d794619ecd6ff14a3f5756eff6a09).
9. I’ve fixed the typo’s I could find using `devtools::spell_check()`. You’ll note that some still appear (like api, aqs), but these are arising from a website link. See [commit](https://github.com/GLOrozcoM/epair/commit/26c21d37e499bc27e0ec75798a2a29fe6979a58b).

Thank you for all your help! I’ll be posting more updates as we make more significant progress.
