### 2021-01-15 #418 @maelle
Thanks for the update and congrats on all the progress @GLOrozcoM & colleagues!
Regarding 6 (and other points), feel free to ask me any question. :slightly_smiling_face:
With spelling you could add a WORDLIST cf https://docs.ropensci.org/spelling/reference/wordlist.html
