### 2021-02-07 #418 @GLOrozcoM
Hi @maelle, thanks for the congrats and waiting!

I'll start with the small things:
1. Per your suggestion, I made a word list. See [commit](https://github.com/GLOrozcoM/epair/commit/045089c3124097a425c030a9de812c2ab845273c).
2. I added a basic R CMD check to the repo to get the ball rolling on CI. See [commit](https://github.com/GLOrozcoM/epair/commit/32dfa7692e495fc6b12ac2d0ad28d664bb34198c). I used [this page](https://ropenscilabs.github.io/actions_sandbox/packageci.html) as a starting point, and will look to add more CI checks as we develop the pkgdown website. I'm not certain if there are other standard checks I should look to incorporate here. Do you have any recommendations?
3. These weren't pointed out by you, but I noticed these and wanted to make sure they were fixed. The PDF manual needed to reflect updates from DESCRIPTION file ([commit](https://github.com/GLOrozcoM/epair/commit/f6d38037b79e42ecd5f7fd9631bd4db79293815e)), and I included a .RBuildignore file ([commit](https://github.com/GLOrozcoM/epair/commit/cd82acda7cad26e8272b7502554f1d84a52d135f)).

Bigger things:
1. Given that `epair` uses `httr` I thought it appropriate to use `httptest` as the package to use for API tests. Your book has been very helpful! I intend on using [this portion](https://books.ropensci.org/http-testing/graceful.html) as a road map for features testing should cover. My impression is some features may be more important than others.  Can you offer some input on what should be prioritized? My intention is to cover them all, but in the short term, I want to make sure I work on the most salient ones.
2. I'm using `httptest` and the `httptest::with_mock_dir()` function to create responses and store these for later. As of now, I'm tweaking these responses to remove authentication. My impression is these should be tweaked to contain the minimum amount of info needed to run the test. Would you say that's the best practice around this?

Thanks in advance for your feedback!
