### 2021-02-08 #418 @maelle
:wave: @GLOrozcoM! Thanks for the update.

Small things :baby_chick:
1. Awesome!
2. Note that after the package is onboarded, the pkgdown website will be built automatically. In the meantime I'd recommend using [`usethis::use_github_action("pkgdown")`](https://usethis.r-lib.org/reference/use_pkgdown.html). You might also like adding an action or steps for testing coverage. From looking at https://github.com/r-lib/actions/tree/master/examples I think the call would be `usethis::use_github_actions("test-coverage")`. Later in my answer I'll also mention the idea to add a scheduled workflow that'd run a R CMD check without cached API responses. (cf https://books.ropensci.org/http-testing/real-requests-chapter.html)
3. :+1:

Bigger things :chicken:
1. Just to clarify you could have used vcr too but yes it will work with httptest! The gracefulness chapter you mention is more about good practice other than testing for API packages. For testing the idea is to get a coverage of at least 75% https://devguide.ropensci.org/building.html#testing We don't have particular requirements around HTTP testing yet but I'd recommend using cached responses (so `httptest::with_mock_dir()` and friends) but having a GitHub Actions workflow, e.g. a scheduled one, with real requests.
2. Why do you need to tweak the responses to remove authentication, where are secrets in the responses? Have you seen https://enpiar.com/r/httptest/articles/redacting.html? Regarding tweaking cached responses to make them minimal, there are two schools of thought on this and you share the view of httptest maintainer which is good given you use httptest. :grin: You could keep the cached responses as they are (tweaking the requests in tests to have requests that don't return tons of stuff) as long as they don't make the package too big for CRAN; or you could indeed tweak them. In https://books.ropensci.org/http-testing/mocking-pkgs-comparison.html#recording-playing-back I have put a box with quotes from Neal Richardson _" In the interest of minimally reproducible examples, and of making tests readable, it often makes sense to take an actual API response and delete a lot of its content, or even to fabricate one entirely."_
