### 2021-02-09 #418 @GLOrozcoM
Great! I'll incorporate your feedback. With regards to tweaking because of authentication, I hadn't seen the enpiar article you mentioned. Thank you for pointing it out! I simply had api key and email being returned in my header within the json file being generated. In any case, I'll make sure to redact carefully here. I'll update again as more gets added.
