### 2021-02-25 #418 @maelle
@GLOrozcoM a small remark as I'm gathering info for the next rOpenSci newsletter via the GitHub API. Your GitHub profile doesn't indicate any name, if you add one it'll be picked up and potentially look more informative (you can also add an URL to a webpage of yours to your GitHub profile for bonus points :grin:).
