### 2021-03-07 #418 @GLOrozcoM
I'm running into an issue I'm hoping has a simple solution @maelle.

I'm testing functions with `httptest::with_mock_dir()` by first using real authentication, then cutting out the authentication info from the json files created. My understanding is that after running with real authentication once, testing will then only happen on the generated file.

So I tried testing by removing my real authentication from my environment (with the expectation that the generated file will be used for testing), but then I get errors about not having my authentication setup. I first saw this with a simple `devtools::test()` command.

For context, `epair` uses `options()` to set up a defined email and api key pair in making requests.

Any suggestions?
