### 2021-03-08 #418 @maelle
Do you use https://enpiar.com/r/httptest/articles/redacting.html?

You might need to trick your package into believing there's authentication i.e. set up some fake authentication when there are mock files. This is what I did with the API key in https://books.ropensci.org/http-testing/httptest.html So in a test helper file (https://blog.r-hub.io/2020/11/18/testthat-utility-belt/), depending on the presence of a mock dir, you'd set up fake authentication credentials with `options()`.

Speaking of `options()`, it might be more secure to use an environment variable. See https://docs.ropensci.org/opencage/articles/opencage.html#authentication-1 for examples of different strategies (e.g. suggesting the users use the keyring package).
