### 2021-03-21 #418 @GLOrozcoM
Yes, to using `capture_requests()`, but no to using the package level redaction functionality. I'll get on that...

Per your suggestion, I've switched authentication to use environment variables in `.Renviron` (see [commit](https://github.com/GLOrozcoM/epair/commit/a394fe8a50ea9b90018afcff6cbda9821626b9f3)). Seems easier to manage than `options()` as well so thanks for the rec! By placing fake authentication logic in `setup()` in the `tests` directory, my testing accepts situations when there are no real credentials (i.e. no errors raised surrounding a lack of authentication) so that part of the problem looks fixed!

My issue has evolved to ensuring genuine calls to the API aren't made after a mock file has been generated. Even after generating a mock file, `httptest::with_mock_dir()` seems to call the API instead of using the mock file. I modified its `replace` argument to `FALSE` in case it kept generating a new file, but this didn't solve anything. I'll keep troubleshooting this, but if you have further suggestions about how to ensure usage of mock files for responses, I'd love to hear them!
