### 2021-03-22 #418 @maelle
is there anything random in what the API receives? Or does the request somehow send the current time?

I don't see the mock files in https://github.com/GLOrozcoM/epair
