### 2021-03-23 #418 @GLOrozcoM
Hmmm. I don't think there's anything random in what's being received. I'll look into the current time for request being sent.

Just uploaded the mock files along with test logic in in this [commit](https://github.com/GLOrozcoM/epair/commit/13c7d14cb6ad6fe827a98e350eb2e7b2ad3b9615). I wanted to wait until I had completely pruned the mock files, but maybe this will help troubleshoot.
