### 2021-03-23 #418 @maelle
Some notes

* Did you edit the mock files by hand or do you now use httptest's support of amending? I got a bit lost.
* I think it'd be best not to use spaces in `httptest::with_mock_dir()` but that is not the source of your problems.
* I see that you e.g. have a JSON file called `countiesByState-3a466f.json` and given the error message when running tests with WiFi off, httptest was expecting `countiesByState-dba3af.json`. In the docs I read "Query parameters and request bodies are incorporated into the file path by hashing". So my best guess is that you created the JSON file with your actual key as query parameter (sounds logical), it was hashed, but when I run the tests a fake key is used so when it gets hashed it does not match with the mock file. I think it might be worth reading https://enpiar.com/r/httptest/articles/redacting.html and then if nothing adds up opening an issue over in the httptest repo.
