### 2021-03-23 #418 @GLOrozcoM
Thanks for the feedback! I'll go back to studying the enpiar article you suggested.

I did indeed edit the files by hand (the headers in responses where the an api key and email is returned). In any case, I'll pursue your suggested line of questioning.
