### 2021-05-09 #418 @GLOrozcoM
Hi @maelle!

Thank you for your patience! We're working out the kinks with pkgdown for documentation and are aiming to have that live by the end of this week (I'll post commits when ready). We're also starting our caching system, and I'm still wrangling our testing together.

We're setting milestones for ourselves in terms of what to get done, and hoping we can completely address the concerns you raised by the latest, June 10.
