### 2021-05-16 #418 @GLOrozcoM
Hi @maelle, some progress!

- For testing, I found out the issue (or corrected the problematic symptom of it) with my use of `httptest`. I changed the name of the file to be that of the expected and could then run the tests correctly. I also redacted, and made sure the mock files were only a few lines in length. You can see the commit [here](https://github.com/GLOrozcoM/epair/commit/9094c1e71a32618e98602bf2944657b5c4632940).

- I added your recommended `test-coverage` action [here](https://github.com/GLOrozcoM/epair/commit/e14f37edb4e9f576c55e341c176ea9ec1a1afdd4). It looks like we're still at around 50% for coverage, partially because of the scraping module (see below). I added a few more tests [here](https://github.com/GLOrozcoM/epair/commit/e7fe92feea40991ef6c3a2a4483adbb42b3e15a4), and will continue to add more robust ones over time.

- One issue I'm thinking about right now is whether to keep the scraping module. `epair` performs scraping to make certain objects that get placed in the data folder as .RData types. My thinking was to keep the scraping logic so it would be clear how the objects were created. I'm not sure now though, if it makes sense to expose this logic as part of the package now since these are very specific to the current API version, and they are already private methods. I'm considering removing this logic from the package, documenting this logic separately in an article in our `pkgdown` site, and updating it when the API changes significantly. Do you have any feedback about this?

- As mentioned in my previous update, I was hoping to make the `pkgdown` version of this site go live, but we're still not ready yet. I'll post more as we get there!
