### 2021-05-17 #418 @maelle
:wave: @GLOrozcoM, thanks for the update!

I think you could put the script in `inst/extdata` cf https://r-pkgs.org/data.html#data-extdata and https://usethis.r-lib.org/reference/use_data.html
