### 2021-06-08 #418 @GLOrozcoM
Hi @maelle! Moved these as you suggested above to `inst/extdata`  [here](https://github.com/GLOrozcoM/epair/commit/5e8d4474a9e4edcaeaada3512857d36c2419a44e).  Incidentally, this boosted our testing coverage, but since we're about to add caching, I wouldn't say we're quite finished yet with testing. Also, we plan on setting up a GitHub scheduled workflow for testing real requests.

We were hoping to complete caching and a `pkgdown` implementation of our tutorial for `epair` by June 10, but we're just not ready yet. We appreciate your patience! I'll continue to post updates here.
