### 2021-07-10 #418 @GLOrozcoM
Hey @maelle! I'm happy to say we have a few updates!

One of our awesome new contributors @mmlawrence has added 1) caching to `epair` [see commit](https://github.com/GLOrozcoM/epair/commit/05d93ac7aab1ae62a37c720b4e42166cf8392eb1#diff-23627626b40c81381a9f780d193e1e2e404cab6fe46aecd47578600808d6e8ed) 2) and the initial set of required docs for our `pkgdown` tutorial/reference page [see commit](https://github.com/GLOrozcoM/epair/commit/50af26c469c73db237fa331a975fc5470df8271c).

These changes are still on our `dev` branch, and we're currently working on merging changes to our production branch. Once these are complete, I'll re update here.
