### 2021-07-25 #418 @GLOrozcoM
@maelle groovy news!

- We've merged caching along with testing for it into our production branch! Our latest commit should show completely passing tests with a coverage of around 80%.
- We also have our `pkgdown` documentation up and visible [here](https://glorozcom.github.io/epair/). This is merely the bare bones to hold reference, and we intend on fleshing it out to be useful and be esthetic soon.

As we look to improve, my understanding is perhaps we are close to being ready for review? We're currently planning on tackling overall improvement to `pkgdown` site and regularly scheduled 'real' API test calls to complement our mock tests. Could you advise on pending reqs before `epair` is ready for review?
