### 2021-07-26 #418 @maelle
:wave: @GLOrozcoM awesome! Yes there are no other pending req. As the websites of onboarded packages are then built with rOpenSci pkgdown template, no need on _styling_ the pkgdown website btw.

I'll start looking for reviewers. Note that I'll be away from keyboard about two weeks from next week.
