### 2021-07-31 #418 @mmlawrence
Hi @maelle! I'll be taking over most of the updates, so for now we are roughly expecting to have most of the content changes we aim to make on the pkgdown website by mid next week (August 3rd or 4th) and depending how smoothly the scheduled API test calls go, that will be done latest by August 11th. If anything changes I'll let you know.
