### 2021-10-03 #418 @KevCaz
- **Briefly describe any working relationship you have (had) with the package authors.** NONE

  ==> There are some minor issues, see below.
- [ ] **Examples** (that run successfully locally) for all exported functions
 ==> There are examples, all are wrapped in `\dontrun{}` (which is ok when they call the web API) so they are not checked and some actually do not work.
  ==> I would add a CONTRIBUTING guideline. This could help potential contributors to have a better idea about how they should handle the tests locally.

  ==> I would recommend the authors to go over the [packaging guide](https://devguide.ropensci.org/building.html) as  there are guidelines that are not currently followed (see below).

Estimated hours spent reviewing: 5h30

---

This is a useful collection of tool to help retrieve data from the Environmental Protection Agency (EPA) API. The code is rather clear and well documented, the authors have made important efforts to explain how the user can create a custom call to the API. Further, API responses can be cached (via `R.cache`) which I think is a useful feature.

IMHO, the authors can go even further and add some feature/functions that would make the package even more useful (see section "Main functionality"). There is also several "not so major"/minor issues that need to be handled too (see below).

#### Main Functionality

In the current version of the package there are functions to create and custom a query (function in `R/call_construction.R`), functions to perform the query and retrieve the response (in `R/call_placement`), functions to transform the data (in `R/data_transforms.R`) and all the data required to do the call (e.g. endpoint names) are available as datasets. I think this is *great* and it is actually what is needed to turn R into a client for the targeted API, this is the **core** of the package.

What the authors should do now (IMHO) is to use this set of functions to create a nice package API, e.g. functions that will do what the users need, automating as many steps as possible! To do so, most of steps required to retrieve the data should happen "behind the scenes". Note that I'm not saying the package should be a black box (the code will remain available), but rather that I would try to make the package more user-friendly.

One way to do this is to start by creating a function that would be flexible to perform any query. The authors could keep all the current functions, making most of them "internals" and then I would create one function that would call them all, so that the user can give a few arguments and obtain a dataset already formatted. Below is a prototype (including some pseudo-code) to illustrate what I'm envisioning.

```R
epa_get_generic <- function(enpoint, by, variable, cached = TRUE, cache_directory = "/cache") {
  # step 1 use the arg + dataset to build the call
  call <- create.call(enpoint, by, variable)
  # step 2
  dataset <- perform.call(call, cached, cache_directory)
  return(dataset)
}
```

Such a function could then be leveraged to create a set of custom functions to access specific data. For instance for the endpoint `dailyData`, I would create another function, say `epa_get_daily()` that would only retrieve data for this endpoint.

```R
epa_get_daily <- function( by, variable, cached = TRUE, cache_directory = "/cache") {
  # call epa_generic() with the dailyData endpoint
  epa_generic("dailyData", by, variable, cached, cache_directory)
}
```

Note that doing so would also document the kind of data that could be retrieved with `epair`. Note I have named the functions above following https://devguide.ropensci.org/building.html#function-and-argument-naming.

### Check

When I run `devtools::check()`, I do have several errors for the tests (see the section "tests") and several warnings and notes that must be addressed. I detail some of these issues is the sections that follow.

```R
❯ checking dependencies in R code ... WARNING
  '::' or ':::' import not declared from: ‘stringr’

❯ checking Rd \usage sections ... WARNING
  Documented arguments not in \usage in documentation object 'add.variables':
    ‘name’

  Documented arguments not in \usage in documentation object 'non.cached.perform.call':
    ‘name’

  Undocumented arguments in documentation object 'perform.call'
    ‘directory’
  Documented arguments not in \usage in documentation object 'perform.call':
    ‘cache_directory’

  Documented arguments not in \usage in documentation object 'perform.call.raw':
    ‘name’

  Functions with \usage entries need to have the appropriate \alias
  entries, and all their arguments documented.
  The \usage entries must correspond to syntactically valid R code.
  See chapter ‘Writing R documentation files’ in the ‘Writing R
  Extensions’ manual.

❯ checking files in ‘vignettes’ ... WARNING
  Files in the 'vignettes' directory but no files in 'inst/doc':
    ‘01_ananomyEPARequest.Rmd’, ‘02_ClarificationTerm.Rmd’,
      ‘03_Installation.Rmd’, ‘04_setup.Rmd’,
      ‘05_making-your-first-call.Rmd’, ‘06_make-a-more-complex-call.Rmd’,
      ‘07_wrap-up.Rmd’, ‘08_Endpoints.Rmd’, ‘09_Services.Rmd’,
      ‘cache/liststates’, ‘epair.Rmd’,
      ‘img/alaska_example/classes_services.png’,
      ‘img/alaska_example/daily_sum.png’, ‘img/alaska_example/dates.png’,
      ‘img/alaska_example/filter.png’,
      ‘img/alaska_example/param_groups.png’,
      ‘img/alaska_example/states.png’,
      ‘img/complex_call/counties_by_state.png’,
      ‘img/endpoints/daily.png’, ‘img/endpoints/endpoint.png’,
      ‘img/endpoints/filter.png’, ‘img/figapi1.png’, ‘img/figapi2.png’,
      ‘img/figapi3.png’, ‘img/list_description.png’,
      ‘img/making_first_call/list_description.png’,
      ‘img/making_first_call/list_filters.png’,
      ‘img/making_first_call/service_selection.png’,
      ‘img/service_selection.png’, ‘img/services/daily_description.png’,
      ‘img/services/monitor_filters.png’,
      ‘img/services/sample_description.png’,
      ‘img/variables/all_vars.png’, ‘img/variables/var_county.png’,
      ‘img/variables/var_types.png’
  Package has no Sweave vignette sources and no VignetteBuilder field.
  The following directory looks like a leftover from 'knitr':
    ‘cache’
  Please remove from your package.

❯ checking for unstated dependencies in ‘tests’ ... WARNING
  'library' or 'require' calls not declared from:
    ‘memoise’ ‘remotes’ ‘stringi’ ‘stringr’

❯ checking for portable file names ... NOTE
  Found the following non-portable file paths:
    epair/tests/testthat/Multiple variable cached call elapsed time/aqs.epa.gov/data/api/dailyData/byState-34db3b.json
    epair/tests/testthat/Raw call for multiple variables returns expected result/aqs.epa.gov/data/api/dailyData/byState-67f884.json
    epair/tests/testthat/Save new cached call saves new call/aqs.epa.gov/data/api/dailyData/byState-14d73f.json
    epair/tests/testthat/Raw call for multiple variables returns expected result/aqs.epa.gov/data/api/dailyData

  Tarballs are only required to store paths of up to 100 bytes and cannot
  store those of more than 256 bytes, with restrictions including to 100
  bytes for the final component.
  See section ‘Package structure’ in the ‘Writing R Extensions’ manual.

❯ checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    ‘LICENSE.md’ ‘cache’

```

###  goodpractice

There are some minor comments I get when I ran `goodpractice::gp()`:

```R
R> goodpractice::gp()
[...]

  ✖ write unit tests for all functions, and all
    package code in general. 81% of code lines are covered by
    test cases.

    R/call_construction.R:32:NA
    R/call_construction.R:35:NA
    R/call_placement.R:12:NA
    R/call_placement.R:13:NA
    R/call_placement.R:14:NA

  ✖ use '<-' for assignment instead of '='. '<-' is
    the standard, and R users and developers are used it and
    it is easier to read your code for them if you use '<-'.

    inst/extdata/making_services.R:217:5
    R/call_construction.R:37:18
    tests/testthat/live_test_setup.R:1:11
    tests/testthat/live_test_setup.R:2:13

  ✖ avoid long code lines, it is bad for readability.
    Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80
    characters

    inst/extdata/epa_scrapers.R:1:1
    inst/extdata/epa_scrapers.R:26:1
    inst/extdata/epa_scrapers.R:44:1
    inst/extdata/epa_scrapers.R:107:1
    inst/extdata/making_services.R:1:1

  ✖ avoid sapply(), it is not type safe. It might
    return a vector, or a list, depending on the input data.
    Consider using vapply() instead.

    inst/extdata/making_services.R:175:23

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error
    prone and result 1:0 if the expression on the right hand
    side is zero. Use seq_len() or seq_along() instead.

    inst/extdata/epa_scrapers.R:90:12
    inst/extdata/epa_scrapers.R:97:12
    inst/extdata/making_services.R:18:12
    inst/extdata/making_services.R:113:12
    R/data_transforms.R:103:12

  ✖ fix this R CMD check WARNING: '::' or ':::'
    import not declared from: ‘stringr’ Namespaces in Imports
    field not imported from: ‘rvest’ ‘xml2’ All declared
    Imports should be used.
```

In general, I try to make `gp()` as happy as I can. Here, the only one I would not bother addressing immediately is the code coverage, because 81% of code coverage is good. BTW, the one that I would try to fix ASAP is the last one (`stringr` missing from `Imports` in `DESCRIPTION`) and I would avoid using `:::` in a package.

#### Getting an API key

All the information needed to obtain a key is readily available, but I think you can further help the user by providing a simple function in the package API

```R
get.key <- function(email) {
  url <- paste0("https://aqs.epa.gov/data/api/signup?email=", email)
  browseURL(url)
  invisible(NULL) # the function silently returns NULL
}
```

Also, I think the documentation should be improved to explain where the key and the email should be stored. So far in the vignette `1. Setup` it is not clear how to permanently add the key and email in `Renviron`, but `create.base.call()` returns the following stop message is it is not set up:

> "Make sure you've declared aqs_api_key in your .Renviron!"

Note that it is not always easy to locate this file, so one way to help the user with this is to mention `usethis::edit_r_environ()`.

Two minor comments here:

1. To use R Markdown to build your vignettes, you must edit the `vignette` field in yaml header accodingly, see
https://r-pkgs.org/vignettes.html#vignette-metadata.

2. In some of the vignette, for several images, a local path is used, so the image won't be added correctly. If I am correct, `/Users/madylawrence/epair/vignettes/` should be removed from those paths.

#### Messages

As mentioned in the rOpenSci packaging guide, it is recommended to use `message()` or `warning()`, instead of `print()` or `cat()`. I would edit `is.API.running()` and use `message()` rather than `print()`. Also, I think this function should return a logical:  `TRUE` if the API is running, `FALSE` otherwise.

There are some mismatches between the argument names in the function and the rgument names in the documentation. From `devtools::check`:

```R
❯ checking Rd \usage sections ... WARNING
  Documented arguments not in \usage in documentation object 'add.variables':
    ‘name’

  Documented arguments not in \usage in documentation object 'non.cached.perform.call':
    ‘name’

  Undocumented arguments in documentation object 'perform.call'
    ‘directory’
  Documented arguments not in \usage in documentation object 'perform.call':
    ‘cache_directory’

  Documented arguments not in \usage in documentation object 'perform.call.raw':
    ‘name’
```

For instance, in `perform.call()`, the name of the argument in the function should be `cache_directory` to be consistent with the documentation.

Currently, unless I have missed something, for the tests to work, the user have to remove all the folders that contains the JSON files. In my opinion, these folders should not be in the repository. Also note that currently, in some of the `.json`, one can find a email/key pair that seems to be a valid one (e.g. https://github.com/GLOrozcoM/epair/blob/master/tests/testthat/Multiple%20variable%20call/aqs.epa.gov/data/api/dailyData/byState-67f884.json).

Once I removed these folders, all tests passed but one:

```R
Loading epair
Updating epair documentation
Loading epair
Writing NAMESPACE
Writing NAMESPACE
Loading epair
Testing epair
✔ |  OK F W S | Context
✖ |   4 1     | call_construction
──────────────────────────────────────────────────────────────────────────
Failure (test_call_construction.R:13:3): Basic call is setup correctly
`found_call` not equal to `exp_call`.
1/1 mismatches
x[1]: "https://aqs.epa.gov/data/api/list/States?&email=kevin.cazelles@gmai
x[1]: l.com&key=taupegazelle65"
y[1]: "https://aqs.epa.gov/data/api/list/States?&email=fake_aqs_email&key=
y[1]: fake_aqs_key"
──────────────────────────────────────────────────────────────────────────
✔ |  15       | call_placement [11.1 s]
✔ |   4       | data_transforms

══ Results ═══════════════════════════════════════════════════════════════

```

Also, while I think the live testing of the API is a good idea, I would rather use your package to do this. So in `testing_live_api.yml` I would install your package and run the tests. The main reason for this is that all the packages declared in `test/testthat/live_test_setup.R` should be added to `Suggests` which is not really necessary here for what you need. This may be slightly more technical and I may not be right, but I can at least say that this will be a problem for a submission to the CRAN.

The first link does not work: https://aqs.epa.gov/aqsweb/documents/data_api => https://aqs.epa.gov/aqsweb/documents/data_api.html

#### Old code

In the documentation, I have found some code that seems to be coming from older versions of the package.  For instance, `get.services()`
is used in several examples, but is no longer found in the package.

In the vignette `Make a more complex call`, in the section "Adding multiple variables", the authors wrote

> Our second option is to create a list that has the appropriate API variable names declared for parameter codes.

But in the current version of the package, only the second option works.

---

Please don't hesitate to contact/ping me if some of the above is not clear enough.
