### 2021-10-04 #418 @maelle
Thanks a lot @KevCaz for your thorough review!

Two comments

* In opencage the docs recommend the users use keyring to keep their secrets: https://docs.ropensci.org/opencage/articles/opencage.html#setup-1
* Regarding JSON in tests, why do tests fail for other contributors @GLOrozcoM? https://enpiar.com/r/httptest/articles/redacting.html might be relevant.
