### 2021-10-04 #418 @GLOrozcoM
@KevCaz - amazing! Thank you for the rigorous feedback. Some of us are traveling and will only be able to start addressing these in about a week. You can be sure, however, we'll implement your suggested changes and will update here once completed.

@maelle - noted for 1). Not sure why tests would fail for other contributors...but will get on this and make sure they can work for other contributors.
