### 2021-10-04 #418 @stefaniebutland
Hi @ogilani. No rush as you mentioned folks are traveling. I'm rOpenSci's Community Manager. I'd like to invite your co-authors to rOpenSci Slack (sent yours already) but have not been able to find their emails. If this is of interest, please send them to me at stefanie@ropensci.org.
