### 2021-10-04 #418 @adithirgis
- **Briefly describe any working relationship you have (had) with the package authors.** Not Applicable

---

Apologies for the late review. Congratulations on developing this useful package and thank you for submitting it. I found this package really helpful for data retrieval using the API provided by Environmental Protection Agency. The whole package is very well documented with vignettes and app for a quick walk through. Apologies for redundant comments, most of these are +1 to @KevCaz's points.

I used RStudio Cloud to check this package, `sessionInfo()`
Running under: Ubuntu 20.04.3 LTS

BLAS:   /usr/lib/x86_64-linux-gnu/atlas/libblas.so.3.10.3
LAPACK: /usr/lib/x86_64-linux-gnu/atlas/liblapack.so.3.10.3

 [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8
 [4] LC_COLLATE=C.UTF-8     LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8
 [7] LC_PAPER=C.UTF-8       LC_NAME=C              LC_ADDRESS=C
[10] LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C

[1] devtools_2.4.2 usethis_2.0.1  magrittr_2.0.1

 [1] xmlparsedata_1.0.5 compiler_4.1.0     prettyunits_1.1.1
 [4] remotes_2.4.1      tools_4.1.0        testthat_3.0.4
 [7] digest_0.6.28      pkgbuild_1.2.0     pkgload_1.2.2
[10] evaluate_0.14      jsonlite_1.7.2     memoise_2.0.0
[13] praise_1.0.0       lifecycle_1.0.1    rlang_0.4.11
[16] rex_1.2.0          whoami_1.3.0       rstudioapi_0.13
[19] cli_3.0.1          yaml_2.2.1         curl_4.3.2
[22] xfun_0.26          xopen_1.0.0        cyclocomp_1.1.0
[25] fastmap_1.1.0      knitr_1.36         pkgreviewr_0.2.0
[28] httr_1.4.2         withr_2.4.2        xml2_1.3.2
[31] desc_1.4.0         fs_1.5.0           rprojroot_2.0.2
[34] glue_1.4.2         R6_2.5.1           processx_3.5.2
[37] rcmdcheck_1.4.0    rmarkdown_2.11     sessioninfo_1.1.1
[40] callr_3.7.0        purrr_0.3.4        lintr_2.0.1
[43] covr_3.5.1         htmltools_0.5.2    clisymbols_1.2.0
[46] ps_1.6.0           ellipsis_0.3.2     goodpractice_1.0.2
[49] lazyeval_0.2.2     cachem_1.0.6       crayon_1.4.1
```

My review starts here -

- Installing `epair` on my local system was a little difficult due to the `.json` files as pointed out by others as well. Is it possible to fix that?

If this helps, while installing in RStudio cloud `devtools::install_github("GLOrozcoM/epair", dependencies = T, build_vignettes = T)` these are the errors I got.
Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     ‘epair/tests/testthat/Multiple variable cached call elapsed time/aqs.epa.gov/data/api/dailyData/byState-34db3b.json’
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     ‘epair/tests/testthat/Raw call for multiple variables returns expected result/aqs.epa.gov/data/api/dailyData/’
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     ‘epair/tests/testthat/Raw call for multiple variables returns expected result/aqs.epa.gov/data/api/dailyData/byState-67f884.json’
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     ‘epair/tests/testthat/Save new cached call saves new call/aqs.epa.gov/data/api/dailyData/byState-14d73f.json’
```
#### Package Check
Using the super useful function `devtools::check()`, I also got similar warnings as shown below.
The quick fix for the `LICENSE.md` file at the top - could be https://stackoverflow.com/questions/48955103/non-standard-file-directory-found-at-top-level-readme-rmd-persists-even-after.
Also I think `:::` is not recommended in a package.

> checking dependencies in R code ... WARNING
  '::' or ':::' import not declared from: ‘stringr’

> checking for code/documentation mismatches ... WARNING
  Codoc mismatches from documentation object 'clear.all.cached':
  clear.all.cached
    Code: function(directory = "/cache")
    Docs: function(directory = "~/epair/cache")
    Mismatches in argument default values:
      Name: 'directory' Code: "/cache" Docs: "~/epair/cache"

  Codoc mismatches from documentation object 'clear.cached':
  clear.cached
    Code: function(endpoint, variables = list(), directory = "/cache")
    Docs: function(endpoint, variables = list(), directory =
                   "~/epair/cache")
    Mismatches in argument default values:
      Name: 'directory' Code: "/cache" Docs: "~/epair/cache"

  Codoc mismatches from documentation object 'list.cached.data':
  list.cached.data
    Code: function(directory = "/cache")
    Docs: function(directory = "~/epair/cache")
    Mismatches in argument default values:
      Name: 'directory' Code: "/cache" Docs: "~/epair/cache"

  Codoc mismatches from documentation object 'perform.call':
  perform.call
    Code: function(endpoint, variables = list(), cached = TRUE, directory
                   = "/cache")
    Docs: function(endpoint, variables = list(), directory =
                   "~/epair/cache", cached = TRUE)
    Mismatches in argument names:
      Position: 3 Code: cached Docs: directory
      Position: 4 Code: directory Docs: cached
    Mismatches in argument default values:
      Name: 'directory' Code: "/cache" Docs: "~/epair/cache"

  Codoc mismatches from documentation object 'retrieve.cached.call':
  retrieve.cached.call
    Code: function(endpoint, variables = list(), directory = "/cache")
    Docs: function(endpoint, variables = list(), user.path)
    Argument names in code not in docs:
      directory
    Argument names in docs not in code:
      user.path
    Mismatches in argument names:
      Position: 3 Code: directory Docs: user.path

  Codoc mismatches from documentation object 'save.new.cached.call':
  save.new.cached.call
    Code: function(endpoint, variables = list(), directory = "/cache")
    Docs: function(endpoint, variables = list(), user.path)
    Argument names in code not in docs:
      directory
    Argument names in docs not in code:
      user.path
    Mismatches in argument names:
      Position: 3 Code: directory Docs: user.path

> checking Rd \usage sections ... WARNING
  Documented arguments not in \usage in documentation object 'add.variables':
    ‘name’

  Documented arguments not in \usage in documentation object
  'non.cached.perform.call':
    ‘name’

  Undocumented arguments in documentation object 'perform.call'
    ‘directory’
  Documented arguments not in \usage in documentation object 'perform.call':
    ‘cache_directory’

  Documented arguments not in \usage in documentation object 'perform.call.raw':
    ‘name’

  Undocumented arguments in documentation object 'retrieve.cached.call'
    ‘user.path’
  Documented arguments not in \usage in documentation object 'retrieve.cached.call':
    ‘directory’

  Undocumented arguments in documentation object 'save.new.cached.call'
    ‘user.path’
  Documented arguments not in \usage in documentation object 'save.new.cached.call':
    ‘directory’

  Functions with \usage entries need to have the appropriate \alias
  entries, and all their arguments documented.
  The \usage entries must correspond to syntactically valid R code.
  See chapter ‘Writing R documentation files’ in the ‘Writing R
  Extensions’ manual.

> checking files in ‘vignettes’ ... WARNING
  Files in the 'vignettes' directory but no files in 'inst/doc':
  ‘01_ananomyEPARequest.Rmd’, ‘02_ClarificationTerm.Rmd’,
      ‘03_Installation.Rmd’, ‘04_setup.Rmd’,
      ‘05_making-your-first-call.Rmd’, ‘06_make-a-more-complex-call.Rmd’,
      ‘07_wrap-up.Rmd’, ‘08_Endpoints.Rmd’, ‘09_Services.Rmd’,
      ‘epair.Rmd’, ‘img/alaska_example/classes_services.png’,
      ‘img/alaska_example/daily_sum.png’, ‘img/alaska_example/dates.png’,
      ‘img/alaska_example/filter.png’,
      ‘img/alaska_example/param_groups.png’,
      ‘img/alaska_example/states.png’,
      ‘img/complex_call/counties_by_state.png’,
      ‘img/endpoints/daily.png’, ‘img/endpoints/endpoint.png’,
      ‘img/endpoints/filter.png’, ‘img/figapi1.png’, ‘img/figapi2.png’,
      ‘img/figapi3.png’, ‘img/list_description.png’,
      ‘img/making_first_call/list_description.png’,
      ‘img/making_first_call/list_filters.png’,
      ‘img/making_first_call/service_selection.png’,
      ‘img/service_selection.png’, ‘img/services/daily_description.png’,
      ‘img/services/monitor_filters.png’,
      ‘img/services/sample_description.png’,
      ‘img/variables/all_vars.png’, ‘img/variables/var_county.png’,
      ‘img/variables/var_types.png’
  Package has no Sweave vignette sources and no VignetteBuilder field.

> checking for unstated dependencies in ‘tests’ ... WARNING
  'library' or 'require' calls not declared from:
    ‘memoise’ ‘remotes’ ‘stringi’ ‘stringr’
    > checking for portable file names ... NOTE
  Found the following non-portable file paths:
    epair/tests/testthat/Multiple variable cached call elapsed time/aqs.epa.gov/data/api/dailyData/byState-34db3b.json
    epair/tests/testthat/Raw call for multiple variables returns expected result/aqs.epa.gov/data/api/dailyData/byState-67f884.json
    epair/tests/testthat/Save new cached call saves new call/aqs.epa.gov/data/api/dailyData/byState-14d73f.json
    epair/tests/testthat/Raw call for multiple variables returns expected result/aqs.epa.gov/data/api/dailyData

  Tarballs are only required to store paths of up to 100 bytes and cannot
  store those of more than 256 bytes, with restrictions including to 100
  bytes for the final component.
  See section ‘Package structure’ in the ‘Writing R Extensions’ manual.

> checking top-level files ... NOTE
  Non-standard file/directory found at top level:
    ‘LICENSE.md’
```

Other than the files which contain actual keys and email ID's, after removing them tests did pass.

#### goodpractice

This is a super amazing package which does a lot of things together. I would like to +1 @KevCaz suggestions about `goodpractice::gp()`. Please do check the package, it comprises of all the good practices in coding like using `lintr` and etc.
My go to reference is here for styling are - http://adv-r.had.co.nz/Style.html and https://devguide.ropensci.org/building.html#code-style. I will still show what I got when I used `goodpractice::gp()`

  ✖ write unit tests for all functions, and all package
    code in general. 81% of code lines are covered by test cases.

    R/call_construction.R:32:NA
    R/call_construction.R:35:NA
    R/call_placement.R:12:NA
    R/call_placement.R:13:NA
    R/call_placement.R:14:NA

  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is
    easier to read your code for them if you use '<-'.

    inst/extdata/making_services.R:217:5
    R/call_construction.R:37:18
    tests/testthat/live_test_setup.R:1:11
    tests/testthat/live_test_setup.R:2:13

    inst/extdata/epa_scrapers.R:1:1
    inst/extdata/epa_scrapers.R:26:1
    inst/extdata/epa_scrapers.R:44:1
    inst/extdata/epa_scrapers.R:107:1
    inst/extdata/making_services.R:1:1

    inst/extdata/making_services.R:175:23

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone
    and result 1:0 if the expression on the right hand side is
    zero. Use seq_len() or seq_along() instead.

    inst/extdata/epa_scrapers.R:90:12
    inst/extdata/epa_scrapers.R:97:12
    inst/extdata/making_services.R:18:12
    inst/extdata/making_services.R:113:12
    R/data_transforms.R:103:12

  ✖ fix this R CMD check WARNING: '::' or ':::' import not
    declared from: ‘stringr’ Namespaces in Imports field not
    imported from: ‘rvest’ ‘xml2’ All declared Imports should be
    used.
  ✖ fix this R CMD check WARNING: Codoc mismatches from
    documentation object 'clear.all.cached': clear.all.cached Code:
    function(directory = "/cache") Docs: function(directory =
    "~/epair/cache") Mismatches in argument default values: Name:
    'directory' Code: "/cache" Docs: "~/epair/cache" Codoc
    mismatches from documentation object 'clear.cached':
    clear.cached Code: function(endpoint, variables = list(),
    directory = "/cache") Docs: function(endpoint, variables =
    list(), directory = "~/epair/cache") Mismatches in argument
    default values: Name: 'directory' Code: "/cache" Docs:
    "~/epair/cache" Codoc mismatches from documentation object
    'list.cached.data': list.cached.data Code: function(directory =
    "/cache") Docs: function(directory = "~/epair/cache")
    Mismatches in argument default values: Name: 'directory' Code:
    "/cache" Docs: "~/epair/cache" Codoc mismatches from
    documentation object 'perform.call': perform.call Code:
    function(endpoint, variables = list(), cached = TRUE, directory
    = "/cache") Docs: function(endpoint, variables = list(),
    directory = "~/epair/cache", cached = TRUE) Mismatches in
    argument names: Position: 3 Code: cached Docs: directory
    Position: 4 Code: directory Docs: cached Mismatches in argument
    default values: Name: 'directory' Code: "/cache" Docs:
    "~/epair/cache" Codoc mismatches from documentation object
    'retrieve.cached.call': retrieve.cached.call Code:
    function(endpoint, variables = list(), directory = "/cache")
    Docs: function(endpoint, variables = list(), user.path)
    Argument names in code not in docs: directory Argument names in
    docs not in code: user.path Mismatches in argument names:
    Position: 3 Code: directory Docs: user.path Codoc mismatches
    from documentation object 'save.new.cached.call':
    save.new.cached.call Code: function(endpoint, variables =
    list(), directory = "/cache") Docs: function(endpoint,
    variables = list(), user.path) Argument names in code not in
    docs: directory Argument names in docs not in code: user.path
    Mismatches in argument names: Position: 3 Code: directory Docs:
    user.path
```
#### Code coverage

I tried rOpenSci's recommended `covr::package_coverage()` for checking the package coverage and it seems to be good at ~ 81%.

#### Extras

- I would update the help file to mention the use of `.Renviron` file and updating it to add the user credentials.

- If you plan to submit it to CRAN, the [DESCRIPTION](https://github.com/GLOrozcoM/epair/blob/master/DESCRIPTION) cannot start with A package or the package. Feel free to ignore this if you are not submitting it to CRAN.

- Also, just like you have mentioned the versions of Imports in the DESCRIPTION file, please add so for Suggests as well.

- I see a few `print()` in the code, as mentioned [here](https://devguide.ropensci.org/building.html), is it possible to change it to `message()` or `warning()`.

- I agree with @KevCaz regarding creating fewer user friendly functions with less variables which makes the relevant calls in the backend and gives the output.

- Also a Contributing file would be helpful for users who would want to contribute. You can try using `usethis::use_tidy_contributing()`.

- I did find the app with the walk through very elegant.

- Not sure, if the index is getting picked up the wrong way or its NULL. I could not get around this problem, let me know if I am doing something wrong or if its the internet issue at my end. I followed the instruction mentioned [here](https://epair.netlify.app/docs/tutorial/3-make-a-more-complex-call/).
While [this](https://epair.netlify.app/docs/sample-query-building/samplesearchdata/) worked perfectly well with different combinations. I had just one follow up question, is there a time constraint or a limitation of data download from EPA, because if I increase the time period I don't get the result.

- Lastly, please feel free to ignore this one, the [styling of functions](https://devguide.ropensci.org/building.html#function-and-argument-naming) recommends function_name while I see function.name.

Just to add in the end - caching data seems to be a very efficient step.

Thank you so much! Please feel free contact me if any clarifications are needed.
Adithi
