### 2021-10-10 #418 @GLOrozcoM
@KevCaz - looking for your thoughts on post removing the JSON files. Totally agree that these seem unnecessary in the repo. If we do this, however, then the majority of tests will fail in CI/CD since these depend on the stored responses within the JSON files. I'm thinking of using the `skip_on_ci` option within `testthat` to skip tests in our CI/CD, but still have them run locally. Does this sound reasonable?
