### 2021-10-11 #418 @KevCaz
@GLOrozcoM, I thought it would be better to remove those, but I would take @maelle's advice on this topic.

I may be wrong about that, but I think what's happening is a mismatch between expected and recorded responses. I'm guessing there is a mismatch between keys and so between calls to API, which leads to a mismatch between recorded JSON and what's expected. I guess a simple solution (not 100% it's the best) would be to keep using`email=fake_aqs_email&key=fake_aqs_key` for all requests (if possible).

Assuming you will keep the JSON I would change their names because I think you are exceeding the path length (see https://cran.r-project.org/doc/manuals/r-release/R-exts.html)

>  In addition, packages are normally distributed as tarballs, and these have a limit on path lengths: for maximal portability 100 bytes.
