### 2021-10-31 #418 @maelle
I'm also wondering whether you know of https://github.com/USEPA/RAQSAPI that has been released on CRAN after your submission. (thanks @jhollist for telling me about it)

How do the two packages compare? E.g. important for the docs see the note about overlap in https://devguide.ropensci.org/building.html#documentation
