### 2021-11-03 #418 @GLOrozcoM
@maelle thank you for asking! We do have an update, but wanted to wait until our branches with fixes were merged and tested internally before posting specific commits. However, what we've worked on and are close to releasing is:
1. Solving testing - allowance for remote mock files, downloading, and shorter path lengths to these.
2. General code and gp/dev checks mentioned by both reviewers above - correct argument names, correct coding style.

We also really appreciate your pointing out `RAQSAPI`. We're currently examining it and will post here regarding relevant comparisons.
