### 2021-11-23 #418 @GLOrozcoM
Hi @maelle, appreciate your patience! We’ve got a load of updates. I’ll point to our latest version in the master branch (https://github.com/GLOrozcoM/epair) that address all of the following.

1. Passing tests remotely and on local using mock files.
    1. Path lengths for mock files cut out.
    2. Use of redact functionality in httptest to avoid any accidental credentials showing.
2. Examples that don’t interact with the API will now be run.
3. `devtools::check()` fully passes without warnings or notes.
4. Some of the issues caught with `goodpractice` check. Still pending here are minor code styling that we’re working on.
5. Function to sign up for the API through `epair`.
6. Vignettes built with rmarkdown and absolute paths removed.
7. `message()` prints to console instead of a `print()`.
8. Documentation mismatches removed.
9. Link works in ReadMe.
10. Removal of old code.

Close to us being down is completing relevant issues found with `goodpractice` checks and adding contribution guidelines.
When completed, I'll post here.

For this cycle, we focused on addressing present major and minor concerns. Our next version of the package intends to implement reviewer suggestion of abstracting further the interaction with the API. Do you regard that as an essential milestone for this package to pass review?
