### 2021-11-23 #418 @GLOrozcoM
Also, @maelle ,  :re your question regarding difference/overlap with `RAQS`.
The aim of both packages is the same - both `epair` and `RAQS` aim to provide an R framework that make calls to the EPA's AQS data mart.
- Two differences, in our opinion, set `epair` apart.
It provides built in caching through R.cache. We believe this will help users setting up data pipeline integrations by avoiding unnecessary live calls to AQS.
- `epair`'s testing framework is built around httptest. This currently provides accurate mocks of real responses, and we intend for this to serve as a foundation for more specific testing as epair grows.
