### 2021-11-25 #418 @maelle
Thanks @GLOrozcoM for the update and congrats on all the progress!

When you are done, could you write responses to reviewers? Then if they have time they can sign off (or I'd check things myself). See the second to last item on https://devguide.ropensci.org/authors-guide.html Thank you!

> Our next version of the package intends to implement reviewer suggestion of abstracting further the interaction with the API. Do you regard that as an essential milestone for this package to pass review?

I'd say this is a change best tackled now indeed, as you get the benefit of the external advice on the changes. :-)
