### 2021-11-28 #418 @GLOrozcoM
@maelle Thank you! Also, thank you for pointing out the author's guide as well. We'll make sure to structure our update in the suggested format.

When complete, I or someone on the team will post an update here.
