### 2021-12-21 #418 @mmlawrence
Thank you so much for both reviews! We appreciate you waiting and we have addressed the suggestions and fixes below with relevant commits for the change.

### Response to @KevCaz:

**Friendly wrapper functions for users!**
- Not finished yet, but latest work is on `master` [here](https://github.com/GLOrozcoM/epair)!

**Warnings and friends found with `devtools::check()`.**
- Tar ball warning, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/92d62fe714c417a1bb7afcf3cd9e7a35ad6357d9)
- General warnings, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/3b1971a585eee44111c4add12eb2327e81ddc001).
- Vignettes and Rmd, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/3b1971a585eee44111c4add12eb2327e81ddc001).

**Warnings and friends found with `goodpractice::gp()`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/2f7188b087fe11f1bad07324a6b7ac1e7cfc3a63)

**Add functionality to signup for an API key through `epair`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/71ebc23a3f76b107a1f8029de8476b31fa47db1a).

**Vignettes and Rmd pow wow.**
- Vignettes and Rmd, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/3b1971a585eee44111c4add12eb2327e81ddc001)
- Removal of local paths, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/3b1971a585eee44111c4add12eb2327e81ddc001).

**Show message if AQS key not setup properly.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).

**Changed `print()` to `warning()` or `message()`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/71ebc23a3f76b107a1f8029de8476b31fa47db1a).

**Documentation mismatch.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/3b1971a585eee44111c4add12eb2327e81ddc001).

**Tests (passing tests without having to remove mock files)**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/273dcb965c342da46c873e874a493e6781e6d304).

**Method for declaring variables not working as documented.**
- Choosing to include only one method for declaring variables, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).

**Broken link when trying to access AQS API website.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/f7dd5663f88638a3d9eae1018617b0e997cdaaf9).

**Old code.**
- Removed old code, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).

**Runnable examples are marked as `\dontrun`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/a2209e561248e20046abb45508d8bf2ca96ddd47).

### Response to @adithirgis:
**Difficulty with installation of `epair`. Storing paths more than 100 bytes is not portable.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/92d62fe714c417a1bb7afcf3cd9e7a35ad6357d9).

**Warnings and friends found with `devtools::check()`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/3b1971a585eee44111c4add12eb2327e81ddc001)
- Removing old code in `extdata`, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).
- Other warnings, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/3b1971a585eee44111c4add12eb2327e81ddc001).

**Tests (passing tests without having to remove mock files)**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/273dcb965c342da46c873e874a493e6781e6d304).

**Warnings shown by `goodpractice::gp()`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/2f7188b087fe11f1bad07324a6b7ac1e7cfc3a63).

**Updating help file to mention use of `.Renviron`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).

**DESCRIPTION not starting with `A package` if submitting to CRAN.**
- Not sure yet if we intend on submitting to CRAN, but changed for the case that we do down the road:
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).

**Add version for Suggests in DESCRIPTION.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).

**Changed `print()` to `warning()` or `message()`.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/71ebc23a3f76b107a1f8029de8476b31fa47db1a).

**Friendly wrapper functions for users!**
- Not finished yet, but latest work is on `master` [here](https://github.com/GLOrozcoM/epair)!

**Contributing file.**
- Addressed in [commit](https://github.com/GLOrozcoM/epair/commit/7762c46836cc9d8cc909f7fbb64fb12436bf807b).

**Time constraints with data downloading.**
- Placed in `Usage notes` in ReadMe, addressed in [commit](https://github.com/GLOrozcoM/epair/commit/0b95580ed5f8ec56bfd5799980bacadbf612f75b).
