### 2021-12-21 #418 @mmlawrence
@maelle We are also in the process of adding a NEWS.md file, but we're wondering if we need to include the changes made after the reviews, or just the basics outlined in the guide since it is still in the submission process? Thanks!
