### 2021-12-26 #418 @adithirgis
Thanks for the update.

I tried to download the package in a Windows 10 system, my `sessionInfo()`

```{r}
> sessionInfo()
R version 4.0.5 (2021-03-31)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 19042)

Matrix products: default

locale:
[1] LC_COLLATE=English_India.1252  LC_CTYPE=English_India.1252    LC_MONETARY=English_India.1252
[4] LC_NUMERIC=C                   LC_TIME=English_India.1252

attached base packages:

 [1] rstudioapi_0.13   magrittr_2.0.1    rcmdcheck_1.3.3   usethis_2.0.1     devtools_2.4.0
 [6] pkgload_1.2.1     R6_2.5.1          rlang_0.4.11      fastmap_1.1.0     fansi_0.4.2
[11] tools_4.0.5       pkgbuild_1.2.0    xopen_1.0.0       sessioninfo_1.1.1 utf8_1.2.1
[16] cli_3.0.1         withr_2.4.1       ellipsis_0.3.2    remotes_2.3.0     digest_0.6.27
[21] rprojroot_2.0.2   tibble_3.1.4      lifecycle_1.0.1   crayon_1.4.1      processx_3.5.0
[26] purrr_0.3.4       callr_3.6.0       vctrs_0.3.8       fs_1.5.0          ps_1.6.0
[31] testthat_3.0.2    memoise_2.0.0     glue_1.4.2        cachem_1.0.4      compiler_4.0.5
[36] pillar_1.6.3      desc_1.3.0        prettyunits_1.1.1 pkgconfig_2.0.3
```

I am confused what happens here when I try to install in my system?
```{r}
The downloaded binary packages are in
	C:\Users\Dell\AppData\Local\Temp\RtmpeIs1CQ\downloaded_packages
√  checking for file 'C:\Users\Dell\AppData\Local\Temp\RtmpeIs1CQ\remotes3de0209dfa7\GLOrozcoM-epair-59b3d10/DESCRIPTION' (639ms)
   Warning in file.copy(pkgname, Tdir, recursive = TRUE, copy.date = TRUE) :
     problem reading directory .\GLOrozcoM-epair-59b3d10\tests\testthat\Retrieve cached call : No such file or directory
    ERROR
   copying to build directory failed
Error: Failed to install 'epair' from GitHub:
  System command 'Rcmd.exe' failed, exit status: 1, stdout + stderr:
E> * checking for file 'C:\Users\Dell\AppData\Local\Temp\RtmpeIs1CQ\remotes3de0209dfa7\GLOrozcoM-epair-59b3d10/DESCRIPTION' ... OK
E> Warning in file.copy(pkgname, Tdir, recursive = TRUE, copy.date = TRUE) :
E>   problem reading directory .\GLOrozcoM-epair-59b3d10\tests\testthat\Retrieve cached call : No such file or directory
E>  ERROR
E> copying to build directory failed
In addition: There were 14 warnings (use warnings() to see them)
```
- For `devtools::check()`
I got this now
```{r}
√  checking for detritus in the temp directory

   See
     'C:/Users/Dell/AppData/Local/Temp/RtmpeIs1CQ/epair.Rcheck/00check.log'
   for details.

-- R CMD check results ------------------------------------------------------------- epair 0.1.0 ----
Duration: 2m 10.2s

> checking tests ...
  See below...

> checking data for ASCII and uncompressed saves ... OK
   WARNING
  'qpdf' is needed for checks on size reduction of PDFs

-- Test failures ---------------------------------------------------------------------- testthat ----

> library(testthat)
> library(epair)
>
> test_check("epair")
== Failed tests ================================================================
-- Failure (test_call_placement.R:110:9): A cached call is faster the second time called --
second.time[[3]][1] is not strictly less than first.time[[3]][1]. Difference: 0.01
-- Error (test_call_placement.R:165:9): Retrieve a previously saved cached call --
Error: GET https://aqs.epa.gov/data/api/dailyData/byState?&email=fake_aqs_email&key=fake_aqs_key&state=37&bdate=20200101&edate=20200102&param=44201 (dailyData/byState-7b08b7.json)
Backtrace:
    x
 1. \-epair::perform.call(endpoint, variable.list) test_call_placement.R:165:8
 2.   \-epair::save.new.cached.call(endpoint, variables, full.directory)
 3.     \-epair::non.cached.perform.call(endpoint, variables)
 4.       \-epair:::place.call(call)
 5.         \-httr::GET(url)
 6.           \-httr:::request_perform(req, hu$handle$handle)
 7.             \-httr:::request_fetch(req$output, req$url, handle)
 8.               \-(function (req, handle, refresh) ...
 9.                 \-httptest:::stop_request(req)

[ FAIL 2 | WARN 0 | SKIP 0 | PASS 119 ]
Error: Test failures
Execution halted

1 error x | 1 warning x | 0 notes √
```

- I am sorry if I am confused a little bit but the link here is not working - https://glorozcom.github.io/epair/index.html. I land at the first page and I am not able to navigate efficiently. Also, the url gets added when I use the navbar. I would recommend having the help site set up for people to navigate smoothly.

![image](https://user-images.githubusercontent.com/63432077/147398575-6a5ac55c-90aa-450e-b65c-005fb82088fa.png)

- The `goodpractice::gp()` also worked out, this is what I get
```{r}
It is good practice to

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    vignettes\a06_make-a-more-complex-call.Rmd:42:1
    vignettes\a06_make-a-more-complex-call.Rmd:81:1
    vignettes\a07_wrap-up.Rmd:57:1
    vignettes\a09_Services.Rmd:25:1
    vignettes\a09_Services.Rmd:NA:1
    ... and 8 more lines

  ✖ fix this R CMD check ERROR: Packages required and available but unsuitable
    versions: 'knitr', 'rmarkdown' See section 'The DESCRIPTION file' in the 'Writing R
    Extensions' manual.
```

- Thanks for updating Time constraints with data downloading, Contributing file, changing `print()` to `message()` and changing function names using best practice.

Please let me know if you need any more details. All the best

- Adithi
