### 2022-01-03 #418 @maelle
Thanks @GLOrozcoM for the response, but could you also summarize changes so reviewers can get information without having to click through each commit? (linking commits is still useful though, should they want and have time to look)

Thanks @adithirgis for checking the package again!

And Happy New Year to all. :-)
