### 2022-01-03 #418 @GLOrozcoM
@maelle for sure! Will absolutely make a summary of changes.

@adithirgis thank you for taking a look! Will also get on fixing what you pointed out.

Hoping 2022 is off to a great start for all as well!
