### 2022-01-17 #418 @GLOrozcoM
@adithirgis thanks for waiting! I'm hoping we can make the installation experience smoother.

1. Can you let us know which commands you're running for the first issue you're facing (no such file, directory)? I've tried a couple of ways on mine, but am not succeeding in reproducing this error.
2. For both `devtools::check()` and  `goodpractice::gp()` I'm also not yet able to reproduce the error. We're going to try a few different installation methods to see if we can get these reproduced. I'll update here when we reproduce and fix.
3. We just overhauled `epair` to use upper level functions (yay!), but our documentation has lagged behind because of it. Not being able to access a useful GitHub Pages site is a symptom of this. We are fixing, however, and @mmlawrence will post here when it is ready to be viewed.

I'm posting separately with regards to a summary as recommended by @maelle.
