### 2022-01-17 #418 @GLOrozcoM
@maelle @adithirgis @KevCaz
This regards our summary of changes.
1. Foremost is no more endpoints! Users can now use custom built functions to perform their calls. This is based primarily on @KevCaz recommendation. This was a major change and results in a substantial amount of new service modules (`_service.R` in R folder) each containing their service related functions. We wrote simple tests for each function with `httptest` and manually tested each as we wrote them.
-> Can test by typing `epair::get_` and seeing autocomplete for list of possible functions. Most of these functions will require a start and end date along with a geographical boundary type (like CBSA code or bounding box). For more details, I recommend going with `?epair::get_` for the function you're interested in testing to see exact required params.
2. Beyond the above, all our changes were related to previous feedback which I'm linking to the general summary here (https://github.com/ropensci/software-review/issues/418#issuecomment-977028124) and the commit level link here (https://github.com/ropensci/software-review/issues/418#issuecomment-998375085).

As mentioned above, we're aiming to fix our GitHub Pages site ASAP to document clearly how to use `epair`, but please let us know if you'd like any details about specific functions meanwhile.
