### 2022-01-18 #418 @adithirgis
Hello!

Thanks for the update.

```{r}
sessionInfo()
R version 4.0.5 (2021-03-31)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 19042)

Matrix products: default

locale:
[1] LC_COLLATE=English_India.1252  LC_CTYPE=English_India.1252
[3] LC_MONETARY=English_India.1252 LC_NUMERIC=C
[5] LC_TIME=English_India.1252

attached base packages:

[1] compiler_4.0.5 tools_4.0.5
```

I tried using `remotes::install_github("GLOrozcoM/epair")` and I updated all the packages,
```{r}
√  checking for file 'C:\Users\Dell\AppData\Local\Temp\RtmpKWDHta\remotes30e0270063b\GLOrozcoM-epair-ec338d5/DESCRIPTION' (390ms)
   Warning in file.copy(pkgname, Tdir, recursive = TRUE, copy.date = TRUE) :
     problem reading directory .\GLOrozcoM-epair-ec338d5\tests\testthat\Retrieve cached call : No such file or directory
    ERROR
   copying to build directory failed
Error: Failed to install 'epair' from GitHub:
  System command 'Rcmd.exe' failed, exit status: 1, stdout + stderr:
E> * checking for file 'C:\Users\Dell\AppData\Local\Temp\RtmpKWDHta\remotes30e0270063b\GLOrozcoM-epair-ec338d5/DESCRIPTION' ... OK
E> Warning in file.copy(pkgname, Tdir, recursive = TRUE, copy.date = TRUE) :
E>   problem reading directory .\GLOrozcoM-epair-ec338d5\tests\testthat\Retrieve cached call : No such file or directory
E>  ERROR
E> copying to build directory failed
```
In a fresh RStudio instance I cloned the project. I got this error

```{r}
error: invalid path 'tests/testthat/Retrieve cached call /dailyData/byState-7b08b7.json'
fatal: unable to checkout working tree
warning: Clone succeeded, but checkout failed.
You can inspect what was checked out with 'git status'
and retry with 'git restore --source=HEAD :/'
```
I downloaded the zip folder of project to run `devtools::check()`

```{r}
> checking data for ASCII and uncompressed saves ... OK
   WARNING
  'qpdf' is needed for checks on size reduction of PDFs

-- Test failures ---------------------------------------------------- testthat ----

> library(testthat)
> library(epair)
>
> test_check("epair")
== Failed tests ================================================================
-- Error (test_call_placement.R:165:9): Retrieve a previously saved cached call --
Error: GET https://aqs.epa.gov/data/api/dailyData/byState?&email=fake_aqs_email&key=fake_aqs_key&state=37&bdate=20200101&edate=20200102&param=44201 (dailyData/byState-7b08b7.json)
Backtrace:
    x
 1. \-epair::perform.call(endpoint, variable.list) test_call_placement.R:165:8
 2.   \-epair::save.new.cached.call(endpoint, variables, full.directory)
 3.     \-epair::non.cached.perform.call(endpoint, variables)
 4.       \-epair:::place.call(call)
 5.         \-httr::GET(url)
 6.           \-httr:::request_perform(req, hu$handle$handle)
 7.             \-httr:::request_fetch(req$output, req$url, handle)
 8.               \-(function (req, handle, refresh) ...
 9.                 \-httptest:::stop_request(req)

[ FAIL 1 | WARN 0 | SKIP 0 | PASS 120 ]
Error: Test failures
Execution halted

1 error x | 1 warning x | 0 notes √
```

 and `goodpractice::gp()` to get this

```{r}
── GP epair ───────────────────────────────────────────────────────────────────────

It is good practice to

    many people prefer editor windows that are about 80 characters wide.
    Try make your lines shorter than 80 characters

    vignettes\a06_make-a-more-complex-call.Rmd:42:1
    vignettes\a06_make-a-more-complex-call.Rmd:81:1
    vignettes\a07_wrap-up.Rmd:57:1
    vignettes\a09_Services.Rmd:25:1
    vignettes\a09_Services.Rmd:NA:1
    ... and 8 more lines

  ✖ checking tests ... Running 'testthat.R' ERROR Running the
    tests in 'tests/testthat.R' failed. Last 13 lines of output:
    Backtrace: x 1. \-epair::perform.call(endpoint, variable.list)
    test_call_placement.R:165:8 2.
    \-epair::save.new.cached.call(endpoint, variables, full.directory) 3.
    \-epair::non.cached.perform.call(endpoint, variables) 4.
    \-epair:::place.call(call) 5.  \-httr::GET(url) 6.
    \-httr:::request_perform(req, hu$handle$handle) 7.
    \-httr:::request_fetch(req$output, req$url, handle) 8.  \-(function
    (req, handle, refresh) ... 9.  \-httptest:::stop_request(req) [ FAIL
    1 | WARN 0 | SKIP 0 | PASS 120 ] Error: Test failures Execution
    halted
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems. LaTeX errors
    found:
```

Let me know if this is a system issue.

Regards
Adithi
