### 2022-01-23 #418 @GLOrozcoM
@adithirgis thanks! I believe the issue was indeed sys related. Windows labels a space at the end of a folder name as illegal (https://kb.acronis.com/content/39790), and that test name had  a space at the end of it.

 As a quick fix, I changed the name of the folder and will later update our CI/CD to include Windows systems. Can you recheck on your system and see if this works now? @mmlawrence will be handling the vignettes errors and will post when ready.
