### 2022-02-04 #418 @adithirgis
Hi!

Sorry for the delayed response. Yes! It works perfectly well now. I have also completed my review process. Please do a final check with `devtools::check()` and `goodpractice::gp()`.

Thanks! All the very best. The package is very useful!

Thank you @maelle, let me know what are the further steps.
