### 2022-03-07 #418 @GLOrozcoM
YES! And thank you for your patience @maelle. We fixed our issue with GitHub Pages (https://glorozcom.github.io/epair/) and documentation for our functions are up!

We're finalizing a change to our ReadMe and entries that use the new higher level user functions, and we'll post here as soon as those are ready.
