### 2022-06-05 #418 @GLOrozcoM
@maelle our pending issues were:
1. Updated ReadMe and docs for functions -> Done!
2. Adding in Windows to CI/CD -> Done!
I say `Done!` as they are addressed in this PR: https://github.com/GLOrozcoM/epair/pull/30. We'll merge it after solving the issue below.

I added in an Ubuntu check and this is looks to be the last hurdle. The Ubuntu checks fail a test we don't normally see failing so we're investigating its fix.
