### 2022-06-19 #418 @GLOrozcoM
@maelle we'd love some help! The test mentioned in question is in the `test_list_service.R` module and starts on line 59. This runs correctly on my Mac system and also passes the remote Windows check. I'm not certain what the best way to debug this is since it works perfectly fine when I try recreating the error on my system. Do have any suggestions?

Some of the red crosses will be addressed within the same PR (mismatch in docs vs code), but the codemeta.json file check is new to us - we'll get on adding it in. Also noticed the check caught two functions without examples, we'll add those in.
