### 2022-06-21 #418 @maelle
Do you mean the test fails on Ubuntu? I'll have a look later this week.
