### 2022-06-23 #418 @maelle
A note regarding CI, I think you could run R CMD check on more platforms and I'd recommend to update to v2 of r-lib GitHub Actions cf https://www.tidyverse.org/blog/2022/06/actions-2-0-0/
