### 2022-06-23 #418 @maelle
Now regarding the issue in the test, I think it's because you have two mock directories with nearly the same name excepting the case, which some OS might confuse? `get_all_pqaos() is OK` and `get_all_pqaos() is ok`.

I'd recommend not using spaces in mock directories name but instead think of them like filenames so for instance pqaos-ok. They're not like `test_that()` arguments.

Hope this helps @GLOrozcoM!
