### 2022-07-17 #418 @GLOrozcoM
@maelle your recommendation worked! Removing spaces fixed the failing test. Our recent work was to:

1. Change the R CMD check to use v2 of r-lib [commit here](https://github.com/GLOrozcoM/epair/commit/137bb3cf74f7bcf667cf44ae0943c6779e7c25b3).
2. Remove spaces and simplify mock directory names [commit here](https://github.com/GLOrozcoM/epair/commit/f5a0efb244e3e4a7bd0f68acf4069938c21629ce) and [here](https://github.com/GLOrozcoM/epair/commit/ea1a501e499c785bfb2536f107c6717a058be4a6).
3. Add examples tag to a few functions that were missing it [commit here](https://github.com/GLOrozcoM/epair/commit/958a892fc98511a45526e63a83edbe455ed79e19).
4. Add a codemeta.json file [commit here](https://github.com/GLOrozcoM/epair/commit/4d7526d445a8940e485cfca0ef0f2df5f87b2c3b).

These looked to us like the final problems to address. Are there others we should take also take a look at?
