### 2022-07-20 #418 @ropensci-review-bot
Approved! Thanks @GLOrozcoM for submitting and @adithirgis, @KevCaz for your reviews! :grin:
